requirements.txt
Pygments>=2.17,<3.0
click>=8.0,<9.0
mistune>=3.0,<4.0
python-docx>=1.1,<2.0
requests>=2.31,<3.0
